Moved adapt to common
